Refactor Z3 translator boundary
