+++ docs/content/docs/roadmap.mdx
@@ -72,7 +72,7 @@ cutover. See [research/10](/research/10_translator_soundness) for the trust-chai
 | Isabelle pivot ([#193](https://github.com/HardMax71/spec_to_rest/issues/193)) | Lean -> Isabelle/HOL + `Code_Target_Scala`; universal `soundness` theorem closes with zero `sorry` | shipped |
 | IR canonicalization ([#202](https://github.com/HardMax71/spec_to_rest/issues/202)) | Full input-language IR in Isabelle; `lower :: expr_full => expr option` projection; flat case-class extraction | shipped |
 | Trust dimension ([#205](https://github.com/HardMax71/spec_to_rest/issues/205)) | Per-check `trust: "sound" \| "best-effort"`; new exit code 4 (`ExitCodes.Trust`); JSON `schemaVersion` 1->2 | shipped |
-| Production routing ([#192](https://github.com/HardMax71/spec_to_rest/issues/192)) | `Translator.translateExpr` routes in-subset shapes through `SpecRestGenerated.translate`; hand-written translator only fires on out-of-subset declaration-level expressions | shipped |
+| Production routing ([#192](https://github.com/HardMax71/spec_to_rest/issues/192)) | `ExpressionEncoder.translateCheckedExpr` routes in-subset check bodies through `SpecRestGenerated.translate` and `SmtTermBridge`; the Scala fallback is isolated to declaration-level expressions | shipped |
 | Codegen drift gate ([#231](https://github.com/HardMax71/spec_to_rest/pull/231)) | `isabelle-build.yml` regenerates `SpecRestGenerated.scala` and diffs against the committed copy on every PR | shipped |
 
 </Collapsible>
@@ -171,7 +171,7 @@ Per-feature ledger for the verifier. For the prose explanation of each entry, se
 | Property-based tests for the Z3/Alloy translators (translation determinism on identical IRs, free-identifier resolution against `TranslatorArtifact`, structural invariant preservation between Z3 and Alloy) via `scalacheck-effect-munit` | shipped ([#162](https://github.com/HardMax71/spec_to_rest/issues/162)), `modules/verify/src/test/scala/specrest/verify/TranslatorPropTest.scala` runs 50 random IRs/property over the verified subset from research/10 |
 | Mechanically verified translator soundness in Isabelle/HOL: universal `soundness` theorem in `proofs/isabelle/SpecRest/Soundness.thy` closes with zero `sorry` over the verified subset; `Code_Target_Scala` extracts `translate`/`eval`/`smt_eval` plus the canonical IR ADT to `modules/ir/src/main/scala/specrest/ir/generated/SpecRestGenerated.scala` | shipped via the Isabelle pivot ([#193](https://github.com/HardMax71/spec_to_rest/issues/193)); IR canonicalized in [#202](https://github.com/HardMax71/spec_to_rest/issues/202); CI runs `isabelle build SpecRest` per PR via `.github/workflows/isabelle-build.yml`. Per-run cert emission deleted post-pivot, universal soundness made it vestigial. |
 | Per-check Trust dimension, each `CheckResult` carries `trust: "sound" \| "best-effort"` driven by the extracted Isabelle `lower` projection; best-effort checks short-circuit to `Skipped` (`category=soundness_limitation`) before backend dispatch; new exit code 4 (`ExitCodes.Trust`) | shipped ([#205](https://github.com/HardMax71/spec_to_rest/issues/205)), JSON `schemaVersion` 1->2 |
-| Production verify path routes in-subset checks via the extracted (Isabelle-verified) translator, `Translator.translateExpr` calls `lower(enums, e)`; on `Some` it routes through `SpecRestGenerated.translate >>> SmtTermToZ3`, on `None` falls back to the hand-written translator (used only by declaration-level expressions and out-of-subset shapes). Best-effort checks at the verify level always skip with `soundness_limitation` (no flag) | shipped ([#192](https://github.com/HardMax71/spec_to_rest/issues/192)) |
+| Production verify path routes in-subset checks via the extracted (Isabelle-verified) translator: `translateCheckedExpr` uses `SpecRestGenerated.translate >>> SmtTermBridge` and fails instead of falling back on `None`; `translateDeclarationExpr` is the separate fallback path for declaration-level refinements. Best-effort checks at the verify level always skip with `soundness_limitation` (no flag) | shipped ([#192](https://github.com/HardMax71/spec_to_rest/issues/192)) |
 
 ## Synthesis follow-ups
